requirements.txt
amqp==2.6.1
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.4
decorator==4.4.2
defusedxml==0.7.0rc2
Django==3.1.6
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
django-timezone-field==4.1.1
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.4.0
djoser==1.1.5
flower==0.9.7
future==0.18.2
html5lib==1.1
humanize==3.2.0
idna==2.10
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
kombu==4.6.11
lxml==4.6.2
MarkupSafe==1.1.1
oauthlib==3.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.2
sqlparse==0.4.1
tornado==6.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.0
